backend/requirements.txt
attrs==25.4.0
Automat==25.4.16
blinker==1.9.0
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
constantly==23.10.4
cryptography==46.0.3
cssselect==1.3.0
defusedxml==0.7.1
dnspython==2.8.0
et_xmlfile==2.0.0
filelock==3.20.3
Flask==3.1.2
Flask-SocketIO==5.4.1
eventlet>=0.33.0
google-auth==2.41.1
google-auth-oauthlib==1.2.3
greenlet==3.3.0
gspread==6.2.1
httplib2==0.31.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
itemadapter==0.13.1
itemloaders==1.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
lxml==6.0.2
MarkupSafe==3.0.3
numpy==2.4.1
oauth2client==4.1.3
oauthlib==3.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
parsel==1.10.0
playwright==1.57.0
Protego==0.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyDispatcher==2.0.7
pyee==13.0.0
pymongo==4.16.0
pyOpenSSL==25.3.0
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
queuelib==1.8.0
requests==2.32.5
beautifulsoup4==4.14.3
requests-file==3.0.1
requests-oauthlib==2.0.0
rsa==4.9.1
Scrapy==2.11.2
scrapy-playwright==0.0.44
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
tldextract==5.3.1
Twisted==25.5.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
w3lib==2.3.1
Werkzeug==3.1.5
zope.interface==8.2
python-dotenv==1.0.1
gunicorn==21.2.0
Flask-Login==0.6.3
flask-cors>=4.0.0
google-generativeai>=0.8.0
